+++ packaging/millennium-helpers/arch-unix-install.sh
@@ -27,9 +27,7 @@ _arch_install_unix_helpers() {
     install -m755 "${dispatcher}" "${pkgdir}/usr/bin/${twin}"
   done
 
-  install -d "${pkgdir}/usr/lib/millennium-helpers/lib"
-  install -m644 scripts/common.sh "${pkgdir}/usr/lib/millennium-helpers/common.sh"
-  install -m644 scripts/lib/*.sh "${pkgdir}/usr/lib/millennium-helpers/lib/"
+  install -d "${pkgdir}/usr/lib/millennium-helpers"
 
   install -Dm644 completions/bash/millennium-helpers "${pkgdir}/usr/share/bash-completion/completions/millennium-helpers"
   local script